+++ omniisaacgymenvs/scripts/rlgames_train.py
@@ -34,7 +34,7 @@ import hydra
 import torch
 from omegaconf import DictConfig
 from omniisaacgymenvs.envs.vec_env_rlgames import VecEnvRLGames
-from omniisaacgymenvs.utils.config_utils.path_utils import retrieve_checkpoint_path
+from omniisaacgymenvs.utils.config_utils.path_utils import retrieve_checkpoint_path, get_experience
 from omniisaacgymenvs.utils.hydra_cfg.hydra_utils import *
 from omniisaacgymenvs.utils.hydra_cfg.reformat import omegaconf_to_dict, print_dict
 from omniisaacgymenvs.utils.rlgames.rlgames_utils import RLGPUAlgoObserver, RLGPUEnv
@@ -91,11 +91,16 @@ def parse_hydra_configs(cfg: DictConfig):
         cfg.device_id = local_rank
         cfg.rl_device = f'cuda:{local_rank}'
     enable_viewport = "enable_cameras" in cfg.task.sim and cfg.task.sim.enable_cameras
+
+    # select kit app file
+    experience = get_experience(headless, cfg.enable_livestream, enable_viewport, cfg.kit_app)
+
     env = VecEnvRLGames(
         headless=headless,
         sim_device=cfg.device_id,
         enable_livestream=cfg.enable_livestream,
         enable_viewport=enable_viewport,
+        experience=experience
     )
 
     # ensure checkpoints can be specified as relative paths